Refactor shader resource management by moving registry functions from CommandDispatch to Renderer3D; update related references in OpenGLShader and clean up unused code.
